Split Name::Taxonomy tests out of name_test.rb (#4708)
Moves classification and ancestors/parents tests (32 tests covering
Name.parse_classification, Name.validate_classification, rank helpers,
and all_parents/parents/children/all_children) into
test/models/name/taxonomy_test.rb, matching app/models/name/taxonomy.rb.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
